Former-commit-id: b54f0e023a7ef8906758b3c3076ba49c95698732

+++ ide/src/edu/wustl/cse/lookingglass/media/ImagesToWebmEncoder.java
@@ -40,27 +40,16 @@ public class ImagesToWebmEncoder extends ImagesToFFmpegEncoder {
 	}
 
 	public void mergeAudio() {
-		//		//remove
-		//		try {
-		//			addAudio( new ScheduledAudioStream( new AudioResource( new File( "C:/Users/Matt/Desktop/blorp/blorpa.wav" ) ), 0 ) );
-		//			addAudio( new ScheduledAudioStream( new AudioResource( new File( "C:/Users/Matt/Desktop/blorp/blorpb.wav" ) ), 0 ) );
-		//			addAudio( new ScheduledAudioStream( new AudioResource( new File( "C:/Users/Matt/Desktop/blorp/blorpc.mp3" ) ), 0 ) );
-		//			addAudio( new ScheduledAudioStream( new AudioResource( new File( "C:/Users/Matt/Desktop/blorp/blorpd.ogg" ) ), 0 ) );
-		//			addAudio( new ScheduledAudioStream( new AudioResource( new File( "C:/Users/Matt/Desktop/blorp/blorpe.mp3" ) ), 0 ) );
-		//			addAudio( new ScheduledAudioStream( new AudioResource( new File( "C:/Users/Matt/Desktop/blorp/blorpf.wav" ) ), 0 ) );
-		//		} catch( IOException e ) {
-		//			e.printStackTrace();
-		//		}
-		//		//remove
-		String blah = this.getVideoPath();
-		String newPath = blah.substring( 0, blah.length() - 5 );
-		newPath += "1.webm";
-		RuntimeUtilities.execSilent( this.ffmpegCommand, "-i", this.getVideoPath(), "-i", getSoundTrack().getAbsolutePath(), newPath );
-		System.out.println( "soundTrack: " + getSoundTrack().getAbsolutePath() );
-		File oldFile = new File( this.getVideoPath() );
-		oldFile.delete();
-		File newFile = new File( newPath );
-		newFile.renameTo( new File( this.getVideoPath() ) );
+		if( getSoundTrack() != null ) {
+			String tempVideoPath = this.getVideoPath();
+			String newPath = tempVideoPath.substring( 0, tempVideoPath.length() - 5 );
+			newPath += "1.webm";
+			RuntimeUtilities.execSilent( this.ffmpegCommand, "-i", this.getVideoPath(), "-i", getSoundTrack().getAbsolutePath(), newPath );
+			File oldFile = new File( this.getVideoPath() );
+			oldFile.delete();
+			File newFile = new File( newPath );
+			newFile.renameTo( new File( this.getVideoPath() ) );
+		}
 	}
 
 	public double getLength() {
